[noodle] Need to trim root prefix to apply ignore
- [ ] when walk the directory, the path pass to ignore patterns contains
the root, i.e. `_example/embed/assets/404/a.html` instead of
`404/a.html`, so pattern `404/*.html` won't match ...
- we are just supporting top level ignore file
- [generator] move generate logic to `generate.go` from `pkg.go`
