+++ noodle/_examples/embed/assets/idx/a.hugefile
@@ -1,0 +1,1 @@
+a.hugefile
