+++ noodle/fs_embed_test.go
@@ -1,0 +1,7 @@
+package noodle
+
+import "testing"
+
+func TestGenerateEmbed(t *testing.T) {
+	GenerateEmbed("_examples/embed/assets")
+}
