--- noodle/generate.go
@@ -1,3 +1,0 @@
-package noodle
-
-// generate assets file with flag
